EventValuesの定義を抽象化
PlayerTeleportEventで定義するのではなくPlayerMoveEventに定義
