+++ src/main/java/jp/nlaocs/skriptcatchup/elements/events/EvtTeleport.java
@@ -6,7 +6,6 @@ import ch.njol.skript.lang.Literal;
 import ch.njol.skript.lang.SkriptEvent;
 import ch.njol.skript.lang.SkriptParser;
 import ch.njol.skript.registrations.EventValues;
-import ch.njol.skript.util.Experience;
 import ch.njol.skript.util.Getter;
 import ch.njol.util.coll.CollectionUtils;
 import org.bukkit.Chunk;
@@ -15,7 +14,6 @@ import org.bukkit.block.Block;
 import org.bukkit.entity.Entity;
 import org.bukkit.event.Event;
 import org.bukkit.event.entity.EntityTeleportEvent;
-import org.bukkit.event.player.PlayerExpChangeEvent;
 import org.bukkit.event.player.PlayerTeleportEvent;
 import org.jetbrains.annotations.Nullable;
 
@@ -33,20 +31,6 @@ public class EvtTeleport extends SkriptEvent {
                         "on creeper teleport:"
                 )
                 .since("1.0, 2.9.0 (entity teleport)");
-        EventValues.registerEventValue(PlayerTeleportEvent.class, Location.class, new Getter<Location, PlayerTeleportEvent>() {
-            @Override
-            @Nullable
-            public Location get(PlayerTeleportEvent event) {
-                return event.getTo();
-            }
-        }, EventValues.TIME_NOW);
-        EventValues.registerEventValue(PlayerTeleportEvent.class, Block.class, new Getter<Block, PlayerTeleportEvent>() {
-            @Override
-            @Nullable
-            public Block get(PlayerTeleportEvent event) {
-                return event.getTo().clone().subtract(0, 0.5, 0).getBlock();
-            }
-        }, EventValues.TIME_NOW);
         EventValues.registerEventValue(EntityTeleportEvent.class, Location.class, new Getter<Location, EntityTeleportEvent>() {
             @Override
             public @Nullable Location get(EntityTeleportEvent event) {
